refactor: useEffect 모두 제거 후 hook을 이용하여 비즈니스 로직 분리
