+++ src/components/FormSection/NameFormSection/NameFormSection.tsx
@@ -1,87 +1,20 @@
-import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
 import PaymentsFormTitle from '../../common/PaymentsFormTitle/PaymentsFormTitle';
 import PaymentsInputField from '../../common/PaymentsInputField/PaymentsInputField';
 
-import ERROR_MESSAGE from '../../../constants/errorMessage';
 import OPTION from '../../../constants/option';
-import REGEX from '../../../constants/regex';
 
 import * as Styled from '../FormSection.styled';
+import useNameFormSection from '../../../hook/useNameFormSection';
 
 const PaymentsInputFieldUppercase = styled(PaymentsInputField)`
     text-transform: uppercase;
   `;
 
-const NameFormSection = ({ ...props }) => {
+const NameFormSection = ({ changeName }: NameFormSectionProps) => {
 
-  const { changeName } = props;
-
-  const [inputState, setInputState] = useState<InputState>({
-    value: '',
-    hasError: false,
-    hasFocus: false,
-    isFilled: false,
-  });
-  const [errorMessage, setErrorMessage] = useState('');
-
-  const handleValueChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = e.target.value;
-    const isFilled = newValue.length === OPTION.nameMaxLength;
-
-    if (newValue.length <= OPTION.nameMaxLength && !REGEX.name.test(newValue)) {
-      setInputState({
-        ...inputState,
-        value: newValue.slice(0, newValue.length - 1),
-        hasError: true,
-      });
-      setErrorMessage(ERROR_MESSAGE.onlyEnglish);
-    } else if (newValue.length > OPTION.nameMaxLength) {
-      setInputState({
-        ...inputState,
-        value: newValue.slice(0, OPTION.nameMaxLength),
-        hasError: false,
-      });
-    } else {
-      setInputState({
-        ...inputState,
-        value: newValue,
-        hasError: false,
-        isFilled: isFilled,
-      });
-    }
-  };
-
-  const handleOnFocus = () => {
-    setInputState({
-      ...inputState,
-      hasFocus: true,
-    });
-  };
-
-  const handleOnBlur = () => {
-    setInputState({
-      ...inputState,
-      hasFocus: false,
-    });
-  };
-
-  useEffect(() => {
-    resetErrors();
-  }, [inputState.hasFocus]);
-
-  useEffect(() => {
-    changeName(inputState.value);
-  }, [inputState.value]);
-
-  const resetErrors = () => {
-    setInputState({
-      ...inputState,
-      hasError: false,
-    });
-    setErrorMessage('');
-  };
+  const [inputState, onChange, errorMessage, handleOnFocus, handleOnBlur] = useNameFormSection({ changeName })
 
   return (
     <Styled.FormSection>
@@ -93,9 +26,9 @@ const NameFormSection = ({ ...props }) => {
             className="name-form-section"
             placeholder="FAMILY / GIVEN"
             maxLength={OPTION.nameMaxLength}
-            value={inputState.value}
-            hasError={inputState.hasError}
-            handleValueChange={(e) => handleValueChange(e)}
+            value={inputState[0].value}
+            hasError={inputState[0].hasError}
+            handleValueChange={(e) => onChange(e)}
             handleOnFocus={handleOnFocus}
             handleOnBlur={handleOnBlur}
           />
